requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.1.0
httpx==0.28.1
apscheduler==3.10.4
pydantic==2.11.3
anthropic==0.52.0
sqlalchemy==2.0.40
